refactor: rename builtin components file
